v3: use narrower integer types in the AST and node-indexed tables
`int` is 64-bit on 64-bit targets now, so every AST field and side table
declared `int` silently doubled in size. Nothing in these tables needs more
than 32 bits: node ids index `FlatAst.nodes`, and positions are file-local
byte offsets.

- `token.Pos.offset/end/id` become `i32`, so the `Pos` embedded in every
  node is 16 bytes instead of 32, and `flat.Node` drops from 104 to 88.
- `flat.NodeId` becomes `i32`, halving `FlatAst.children` (one entry per
  AST edge) and every node-id side table, including the checker's
  direct/rewritten parent indexes.
- `token.File.line_offsets`, `FlatAst.file_node_ids`, the checker's
  top-level/synthetic/inactive declaration indexes, `fn_check_costs`,
  `strings_builder_candidates`, `Scope.generations`, the transformer's
  `source_parent_ids`/`fn_scan_costs`, cgen's `top_level_node_ids` and the
  driver's import-scan id lists become `[]i32`.
- The transformer's per-node module/file context tables kept a 24-byte
  string header per node for a handful of distinct spellings; they now keep
  a `u32` id into a small interned table.

Reads promote to `int` implicitly, so consumers are unchanged.

Self-hosting `cmd/v` (macOS/arm64), RSS after each stage:

                  before     after
    parse          530 MB    494 MB
    check         1852 MB   1816 MB
    markused      2907 MB   2861 MB
    transform     3474 MB   3351 MB
    cgen          3007 MB   2905 MB
    peak          5387 MB   5186 MB

Codegen is unchanged: gen2 and gen3 of the self-host emit byte-identical C.

+++ vlib/v3/types/checker_tail.v
@@ -1256,7 +1256,7 @@ fn (tc &TypeChecker) enum_initializer_calls_helper(fn_name string, module_name s
 	target_module := if module_name in ['', 'main'] { '' } else { module_name }
 	mut indexes := tc.top_level_idx.clone()
 	if indexes.len == 0 {
-		indexes = []int{len: tc.a.nodes.len, init: index}
+		indexes = []i32{len: tc.a.nodes.len, init: i32(index)}
 	}
 	mut cur_module := ''
 	for idx in indexes {
